consolidate EPub + Manifest
